refactor(core): shorthands to other files
