+++ packages/core/src/styles/directions.ts
@@ -1,0 +1,1 @@
+export const directions = ['top', 'right', 'bottom', 'left'] as const;
